Tidy file
